Move ask/review command logic and RAG init to llm.rs
